Revert "Rename directory from src to lib"
