+++ document_ja/max_flow.md
@@ -94,7 +94,7 @@ graph.edges
 ## 参考リンク
 
 - 当ライブラリ
-  - [当ライブラリの実装コード max_flow.rb(GitHub)](https://github.com/universato/ac-library-rb/blob/master/lib/max_flow.rb)
+  - [当ライブラリの実装コード max_flow.rb(GitHub)](https://github.com/universato/ac-library-rb/blob/master/src/max_flow.rb)
   - [当ライブラリのテストコード max_flow_test.rb(GitHub)](https://github.com/universato/ac-library-rb/blob/master/test/max_flow_test.rb)
 - 本家ライブラリ
   - 本家のコード
