+++ src/convolution.rb
