+++ src/fenwick_tree.rb
