+++ src/floor_sum.rb
