+++ src/lazy_segtree.rb
