+++ src/lcp_array.rb
