+++ src/scc.rb
