+++ src/segtree.rb
