+++ src/suffix_array.rb
