+++ src/z_algorithm.rb
