+++ test/dsu_test.rb
@@ -3,7 +3,7 @@
 require 'minitest'
 require 'minitest/autorun'
 
-require_relative '../lib/dsu.rb'
+require_relative '../src/dsu.rb'
 
 class UnionFindTest < Minitest::Test
   def test_zero
