+++ test/example/convolution_practice.rb
@@ -3,7 +3,7 @@
 # TLE, but if you replace all @mod with 998244353 and define all method in top level, you may be able to get AC.
 # https://atcoder.jp/contests/practice2/submissions/16655600
 
-require_relative '../../lib/convolution.rb'
+require_relative '../../src/convolution.rb'
 
 _, a, b = $<.map{ |e| e.split.map &:to_i }
