+++ test/example/segtree_practice.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../../lib/segtree.rb'
+require_relative '../../src/segtree.rb'
 
 INF = (1 << 60) - 1
